Fieldset NG: Supports 'overflow' property properly
The 'overflow' property value should be propagated from a FIELDSET to
its anonymous fieldset content box, and its LEGEND should not be
scrolled if the anonymous fieldset content box is scrolled.

This CL consists of two parts:
A) Move out the rendered legend box from the anonymous fieldset content
   box.  The rendered legend should be the first child of the FIELDSET.
B) Propagate 'overflow' property value of a FIELDSET to its anonymous
   fieldset content box, and disable the 'overflow' property for the
   FIELDSET.

A)
The main change is LayoutNGFieldset::AddChild().
Because the rendered legend depends on the order in the children, we
need to reattach FIELDSET's LayoutObject whenever
 - a LEGEND is added to a FIELDSET
   HTMLLegendElement::InsertedInto()
 - a LEGEND is removed from a FIELDSET
   HTMLLegendElement::RemovedFrom()
 - LEGEND's 'position' 'float' properties are changed
   ComputedStyle::NeedsReattachLayoutTree() and
   Element::RecalcOwnStyle()

This CL removes some code to handle the situation that the rendered
legend is a child of the anonymous fieldset content box.
  - LayoutFieldset::FindInFlowLegend()
  - LayoutObject::IsRenderedLegendInternal()
  - GetLayoutObjectForFirstChildNode()
  - GetLayoutObjectForParentNode()
  - GetLayoutObjectForNextSiblingNode()
  - NGBlockNode::PlaceChildrenInLayoutBox()

B)
In HTMLFieldSetElement::CustomStyleForLayoutObject(), we reset
'overflow' property values and store the original values.
LayoutNGFieldset::AddChild() copies the original values to the anonymous
fieldset content box.

Test changes
Update fieldset-content-before-legend.html and
fieldset-generated-content.html so that they are not affected by
rounding errors by fractional paddings and fractional letter widths.

Many testcases in NGFieldsetLayoutAlgorithmTest now has no anonymous
fieldset content boxes.

Bug: 875235, 786475
Change-Id: Idc54069a75b2fba46897267fccd35c44c89b0c1f