Remove the exported build of the WebGPU CTS from WPT
Updates the README in WPT to reflect this, and Chromium's roll script so
that it doesn't roll into WPT.

And, unrelatedly, removes JusSn from the suggested_reviewers as he is
no longer actively working on WebGPU.

Rationale:

These tests test an unstable, constantly changing API. Because of this,
it's not practical for browser implementations to try to sync on a
single version of the test suite which is auto-imported/exported from
WPT. Auto-imports in any browser will necessarily result in test
failures just because the tests have changed to match some API change.

Instead, for the moment, browsers should pin ("vendor") a particular
version of the WebGPU CTS, and may include a built version of it in
their equivalent of "wpt_internal" for automated testing.

Discussion: https://github.com/web-platform-tests/wpt/issues/24699

Bug: None
Change-Id: Iaabd81ef646404ab7f6e8cc19a884b15f5306bca